move index-writer-tests into host
