+++ .github/workflows/ci.yaml
@@ -344,7 +344,6 @@ jobs:
             "card-endpoints-test.ts",
             "card-source-endpoints-test.ts",
             "file-watcher-events-test.ts",
-            "index-writer-test.ts",
             "indexing-test.ts",
             "transpile-test.ts",
             "module-syntax-test.ts",
