+++ packages/realm-server/tests/index.ts
@@ -8,7 +8,6 @@ import './card-dependencies-endpoint-test';
 import './card-endpoints-test';
 import './card-source-endpoints-test';
 import './file-watcher-events-test';
-import './index-writer-test';
 import './indexing-test';
 import './module-syntax-test';
 import './permissions/permission-checker-test';
